refactor: 폴더 위치, 메서드 이름 변경으로 인한 수정 사항 변경
